# src/main/resources/openapi/openapi-geography.yaml
openapi: 3.0.3
info:
  title: Vacapp Geographic Control API
  description: API REST para gestión jerárquica de terrenos ganaderos (Rancho → Sección → Potrero)
  version: 1.0.0
  contact:
    name: Vacapp Team
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/geography/secciones:
    post:
      tags:
        - Secciones
      summary: Crear nueva sección
      operationId: createSeccion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSeccionRequest'
      responses:
        '201':
          description: Sección creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeccionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    
    get:
      tags:
        - Secciones
      summary: Listar secciones del tenant
      operationId: listSecciones
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Lista de secciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SeccionResponse'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/geography/secciones/{id}:
    get:
      tags:
        - Secciones
      summary: Obtener sección por ID
      operationId: getSeccionById
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Sección encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeccionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Secciones
      summary: Actualizar sección
      operationId: updateSeccion
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSeccionRequest'
      responses:
        '200':
          description: Sección actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeccionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Secciones
      summary: Archivar sección
      operationId: archiveSeccion
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Sección archivada exitosamente
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      description: UUID del recurso
      schema:
        type: string
        format: uuid
  schemas:
...
    CreateSeccionRequest:
      type: object
      required:
        - nombre
        - superficie
        - ranchoId
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
          example: "Sección Norte"
        superficie:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 999999999
          example: 20000.00
        ranchoId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        descripcion:
          type: string
          maxLength: 500
    UpdateSeccionRequest:
      type: object
      required:
        - nombre
        - superficie
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
        superficie:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 999999999
        descripcion:
          type: string
          maxLength: 500
    
    # ===== REQUEST DTOs - POTRERO =====
...
    SeccionResponse:
      type: object
      properties:
        seccionId:
          type: string
          format: uuid
        nombre:
          type: string
        superficie:
          type: number
          format: decimal
        superficieDisponible:
          type: number
          format: decimal
        superficieUsada:
          type: number
          format: decimal
        ranchoId:
          type: string
          format: uuid
        descripcion:
          type: string
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 0
          description: Página actual (0-based)
        size:
          type: integer
          minimum: 1
          description: Tamaño de página
        total:
          type: integer
          minimum: 0
          description: Total de elementos
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Petición inválida (validación fallida)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Autenticación requerida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflicto (ej. nombre duplicado)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'